refactor: colocate test files with implementations
Move all test files to sit directly next to their implementation files,
improving project organization and following modern testing conventions.
