--- src/__tests__/placeholder.test.ts
@@ -1,7 +1,0 @@
-import { describe, it, expect } from "vitest";
-
-describe("Test infrastructure", () => {
-  it("should run tests", () => {
-    expect(true).toBe(true);
-  });
-});
