+++ src/lib/loans/plans.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { PLAN_CONFIGS, CURRENT_RATES } from "../plans";
+import { PLAN_CONFIGS, CURRENT_RATES } from "./plans";
 
 describe("PLAN_CONFIGS", () => {
   describe("required fields", () => {
