+++ vitest.config.ts
@@ -6,13 +6,13 @@ export default defineConfig({
   plugins: [react(), tsconfigPaths()],
   test: {
     environment: "jsdom",
-    setupFiles: ["./src/__tests__/setup.ts"],
+    setupFiles: ["./src/test/setup.ts"],
     coverage: {
       provider: "v8",
       reporter: ["text", "html", "lcov"],
       thresholds: { lines: 20, functions: 20, branches: 14, statements: 20 },
       include: ["src/**/*.ts", "src/**/*.tsx"],
-      exclude: ["src/**/*.test.ts", "src/__tests__/**"],
+      exclude: ["src/**/*.test.ts", "src/**/*.test.tsx"],
     },
   },
 });
